ISSUE #2093 - Renamed issueReference -> ticketReference
