+++ frontend/routes/components/messagesList/components/message/components/markdownMessage/ticketReference/index.ts
@@ -15,4 +15,4 @@
  *  along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-export { default as IssueReference } from './issueReference.container';
+export { default as TicketReference } from './ticketReference.container';
