+++ frontend/routes/components/messagesList/components/message/components/markdownMessage/ticketReference/ticketReference.styles.ts
